pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "neurodatahub-cli"
dynamic = ["version"]
description = "A command-line interface for downloading neuroimaging datasets"
authors = [
    {name = "NeuroDataHub Team", email = "contact@neurodatahub.org"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
keywords = ["neuroimaging", "datasets", "cli", "neuroscience", "mri", "brain"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: System :: Archiving",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0",
    "requests>=2.25.0",
    "selenium>=4.0.0",
    "tqdm>=4.60.0",
    "rich>=12.0.0",
    "colorama>=0.4.4",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "isort>=5.10",
    "flake8>=5.0",
    "mypy>=1.0",
    "pre-commit>=3.0.0",
    "bandit>=1.7.0",
    "build>=0.10.0",
    "twine>=4.0.0",
]
performance = [
    "psutil>=5.9.0",
    "aiohttp>=3.8.0",
    "aiofiles>=23.0.0",
]

[project.urls]
Homepage = "https://blackpearl006.github.io/NeuroDataHub/"
Repository = "https://github.com/blackpearl006/neurodatahub-cli"
Documentation = "https://blackpearl006.github.io/NeuroDataHub/"
"Bug Reports" = "https://github.com/blackpearl006/neurodatahub-cli/issues"

[project.scripts]
neurodatahub = "neurodatahub.cli:main"

[tool.setuptools]
packages = ["neurodatahub"]
include-package-data = true

[tool.setuptools.package-data]
neurodatahub = ["data/*.json"]

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "dirty-tag"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=neurodatahub"
testpaths = ["tests"]